requirements.txt
absl-py==0.7.1
astor==0.7.1
backcall==0.1.0
cycler==0.10.0
decorator==4.4.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.20.1
h5py==2.9.0
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
networkx==2.4
numpy==1.16.5
opencv-python==4.1.1.26
pandas==0.23.4
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.1.1
prompt-toolkit==2.0.10
protobuf==3.7.1
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.4
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.1.0
six==1.12.0
tensorboard==1.13.1
tensorflow==2.3.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tqdm==4.36.1
pytorch-nlp==0.5.0
torch-geometric==1.5.0
traitlets==4.3.3
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2